+++ docs/CHANGELOG.md
@@ -1,3 +1,19 @@
+Version 3.21.0-eisop-1 (January ?, 2022)
+-------------------------------
+
+**User-visible changes:**
+
+`DefaultQualifier` supports the new `applyToSubpackages` annotation attribute
+to decide whether a default should also apply to subpackages. To preserve the
+current behavior the default is `true`.
+
+**Implementation details:**
+
+
+**Closed issues:**
+
+
+
 Version 3.21.0 (December 17, 2021)
 -------------------------------